chore: split internal/types/types.go into multiple files
### what

split internal/types/types.go into multiple files

### why

smaller files, per-topic

### testing

n/a

### docs

n/a
